Refactor to an actual library
