+++ src/hooks/use-navigate.js
@@ -1,0 +1,7 @@
+const useNavigate = () => {
+  throw new Error(
+    "useNavigate is removed. Use import { navigate } from 'gatsby' instead"
+  );
+};
+
+export { useNavigate };
