+++ src/lib/context-base.js
@@ -1,0 +1,8 @@
+import { createNamedContext } from "./context-create-named";
+
+let BaseContext = createNamedContext("Base", {
+  baseuri: "/",
+  basepath: "/"
+});
+
+export { BaseContext };
