+++ src/lib/context-create-named.js
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+
+const createNamedContext = (name, defaultValue) => {
+  const Ctx = createContext(defaultValue);
+  Ctx.displayName = name;
+  return Ctx;
+};
+
+export { createNamedContext };
