+++ src/lib/context-location.js
@@ -1,0 +1,5 @@
+import { createNamedContext } from "./context-create-named";
+
+let LocationContext = createNamedContext("Location");
+
+export { LocationContext };
